ST3/requirements.txt
pytorch==1.7.1
transformers==4.0.0
pytorch_lightning==1.2.0
